requirements-dev.txt
openai
requests
regex
tiktoken
azure-cosmos
azure-messaging-webpubsubservice
azure-search-documents
azure-identity
azure-storage-blob
python-dotenv
